% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bcra.R
\name{get_bcra}
\alias{get_bcra}
\title{Datos de la API del BCRA}
\usage{
get_bcra(id_vars, start_date = "2024-01-01", end_date = "2025-12-31")
}
\arguments{
\item{id_vars}{Codigo de la variable a consultar o vector con multiples codigos}

\item{start_date}{Fecha de inicio de la consulta}

\item{end_date}{Fecha de fin de la consulta}
}
\value{
Dataframe con los datos de la consulta
}
\description{
Devuelve los datos de la consulta a la API en formato data frame
#'
}
\examples{

ids <- c(6,14)
fecha_inicio <- '2024-01-01'
fecha_fin <- Sys.Date()
tasas <- get_bcra(id_vars = ids, start_date = fecha_inicio, end_date = fecha_fin)

}
