% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{get_datasets}
\alias{get_datasets}
\title{Series disponibles}
\usage{
get_datasets(pattern = NULL)
}
\arguments{
\item{pattern}{Opcional: palabras claves para filtrar las series. Por defecto muestra todas las disponibles}
}
\value{
Dataframe con las series con su código para poder consultar
}
\description{
Muestra las series disponibles en la API  Principales variables de Estadísticas Monetarias
https://www.bcra.gob.ar/BCRAyVos/catalogo-de-APIs-banco-central.asp
}
\examples{

series_disponibles <- get_datasets('tasas')

}
