% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mercado_cambios.R
\name{get_mercado_cambios}
\alias{get_mercado_cambios}
\title{Datos del Mercado de Cambios}
\usage{
get_mercado_cambios()
}
\value{
Dataframe con los datos del mercado de cambios
}
\description{
Devuelve los datos del Anexo estadístico del informe mensual del mercado de cambios, correspondiente al excel 'Nuevo Anexo', pestaña 'Datos mercado de cambios', en formato data frame
}
\examples{

data <- get_mercado_cambios()
head(data)
}
